+++ packages/reporting/src/request/hash.js
@@ -8,7 +8,7 @@
 
 /* eslint no-param-reassign: 'off' */
 
-import { isIpv4Address } from '../utils/url';
+import { isIpv4Address } from './utils/url';
 import prob from './resources/prob';
 
 export class HashProb {
