+++ packages/reporting/src/request/steps/cookie-context.js
@@ -7,9 +7,9 @@
  */
 
 import { truncatedHash } from '../../md5';
-import pacemaker from '../../utils/pacemaker';
-import { parse } from '../../utils/url';
-import { sameGeneralDomain, getGeneralDomain } from '../../utils/tlds';
+import pacemaker from '../utils/pacemaker';
+import { parse } from '../utils/url';
+import { sameGeneralDomain, getGeneralDomain } from '../utils/tlds';
 
 import { cleanTimestampCache } from '../utils';
