+++ packages/reporting/src/request/steps/token-checker.js
@@ -9,8 +9,7 @@
 /* eslint no-param-reassign: 'off' */
 
 import md5, { truncatedHash } from '../../md5';
-import { tryDecodeURIComponent } from '../../utils/url';
-
+import { tryDecodeURIComponent } from '../utils/url';
 import TokenDomain from '../token-domain';
 import BlockLog from '../block-log';
