+++ packages/reporting/src/request/steps/token-examiner.js
@@ -10,7 +10,7 @@
 
 import * as datetime from '../time';
 import md5, { truncatedHash } from '../../md5';
-import pacemaker from '../../utils/pacemaker';
+import pacemaker from '../utils/pacemaker';
 
 class TokenSet {
   constructor() {
